+++ sdk/deviceprovisioningservices/arm-deviceprovisioningservices/src/iotDpsClient.ts
@@ -58,7 +58,7 @@ export class IotDpsClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-deviceprovisioningservices/5.0.1`;
+    const packageDetails = `azsdk-js-arm-deviceprovisioningservices/6.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -105,7 +105,7 @@ export class IotDpsClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2022-02-05";
+    this.apiVersion = options.apiVersion || "2022-08-07";
     this.operations = new OperationsImpl(this);
     this.dpsCertificate = new DpsCertificateImpl(this);
     this.iotDpsResource = new IotDpsResourceImpl(this);